chore: move widgets to their separate folder
