+++ src/printer/widgets/north-arrow.js
